ref(spans): Remove `span_v2_to_span_v1`
This removes the function `span_v2_to_span_v1` and supporting code, as
it's unused. The `derive_op_for_v2_span` function, which is used in
normalization, is moved to a new `op` module.
