+++ relay-spans/src/lib.rs
@@ -7,15 +7,13 @@
 )]
 
 pub use crate::name::name_for_span;
+pub use crate::op::derive_op_for_v2_span;
 pub use crate::otel_to_sentry_v2::otel_to_sentry_span as otel_to_sentry_span_v2;
-pub use crate::v2_to_v1::{derive_op_for_v2_span, span_v2_to_span_v1};
+pub use crate::v1_to_v2::span_v1_to_span_v2;
 
 pub use opentelemetry_proto::tonic::trace::v1 as otel_trace;
 
 mod name;
+mod op;
 mod otel_to_sentry_v2;
-mod status_codes;
 mod v1_to_v2;
-mod v2_to_v1;
-
-pub use v1_to_v2::span_v1_to_span_v2;
